Restructuring
Because I got lost in what I want to do I decided to make it a little bit less complex and to remove the heavy 3rd database normalization. I will try to focus on the permissions while I am stuck into wanting to use JWT with Config permissions instead of just passing it when calling other methods for different endpoints.