Fixes after teacher's comments
Fixes in the Token handling and message encrypt and decrypt are made after teacher's comments on the code. The backend part seems to be working now.
The logged in user can read their messages, can create new ones. The messages are enrypted in the database but read in plain text. Next step is to create the client part.

Currently all users use the same hardcoded encryption and decryption key.